% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_version_warning}
\alias{check_version_warning}
\title{Check for Newer Version of turtle}
\usage{
check_version_warning(repo = "drhealy013/turtle")
}
\arguments{
\item{repo}{A character string specifying the GitHub repository in the format `"username/repo"`. Default is `"drhealy013/turtle"`.}
}
\value{
A message if an update is available; otherwise, returns invisibly.
}
\description{
Checks whether the installed version of the `turtle` package is older than the latest version available on GitHub.
If so, it suggests updating the package.
}
\examples{
\dontrun{
check_version_warning()
}

}
