% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ttl_files.R
\name{combine_ttl_files}
\alias{combine_ttl_files}
\title{Combine TTL files on Wikibase}
\usage{
combine_ttl_files(
  qids,
  output_file,
  base_url = "https://reprexbase.eu/fu/Special:EntityData/"
)
}
\arguments{
\item{qids}{The QIDs of the RDF datasets in TTL}

\item{output_file}{An output file}

\item{base_url}{The base URL of the Wikibase instance}
}
\value{
A text file with a combined TTL file.
}
\description{
Combine TTL files on Wikibase
}
