% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_diagnostics.R
\name{generate_summary_page}
\alias{generate_summary_page}
\title{Create Summary Page for Diagnostic PDF}
\usage{
generate_summary_page(
  n_models,
  flagged_n,
  diagnostics,
  outlier_threshold,
  pdf_path,
  readme_url
)
}
\arguments{
\item{n_models}{Integer. Total number of models processed.}

\item{flagged_n}{Integer. Number of models exceeding the outlier threshold.}

\item{diagnostics}{Character vector of diagnostics that were run.}

\item{outlier_threshold}{Numeric. Threshold used to flag models based on outlier percentage.}

\item{pdf_path}{String. Path where the PDF is saved.}

\item{readme_url}{String. URL to the online documentation or README.}
}
\value{
A `grid` text grob object to be rendered in the PDF.
}
\description{
Generates a summary page for the diagnostic PDF report. The summary includes a count of models processed, a list of diagnostics performed with plain-language explanations, and a link to further documentation.

This function is used internally by `run_model_diagnostics()` when `save_pdf = TRUE`.
}
\keyword{diagnostics}
\keyword{internal}
