% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getdata}
\alias{getdata}
\title{Get data}
\usage{
getdata(dataset, package)
}
\arguments{
\item{dataset}{The name of the dataset.}

\item{package}{The name of the package.}
}
\description{
Get an internal dataset from an R package.
}
\examples{
getdata("dataset_namespace", package = "dataset") 
}
\keyword{internal}
