% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_assignment_reminder}
\alias{print_assignment_reminder}
\title{Print a Reminder to Assign Function Output}
\usage{
print_assignment_reminder(function_name = "this function")
}
\arguments{
\item{function_name}{A string indicating the name of the function being run (e.g., `"run_linear_models"`).
This will be used in the printed message to personalize the guidance.}
}
\value{
No return value. This function is called for its side effect: printing a message to the console.
}
\description{
This utility function checks whether a function was called without assigning its output to a variable.
If so, it prints a helpful reminder to the user, encouraging them to assign the result so they can access it later.
This is especially useful for guiding users who are less familiar with R.
}
\details{
This function is intended to be called at the end of user-facing functions. It uses `sys.nframe()` to determine
whether the function was called at the top level (i.e., not assigned to a variable). If so, it prints a message
explaining how to assign the result and how to access key components like `$tidy`.
}
\examples{
print_assignment_reminder("run_linear_models")

}
