% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_linear_models.R
\name{run_linear_models}
\alias{run_linear_models}
\title{Fit a Linear or Mixed Effects Model with Optional Modifiers}
\arguments{
\item{data}{Your dataset containing the variables used in the model.}

\item{outcome}{A string specifying the outcome variable. This can be written as: outcomes <- set_names(c("outcome_1", "outcome_2")) and so on for how many outcomes you are investigating}

\item{exposure}{A string specifying the main exposure variable. This can be written as: exposures <- set_names(c("exposure_1", "exposure_2")) and so on for how many exposures you are investigating}

\item{covariates}{A character vector of covariate names to adjust for. Write your covariates as a normal formula: covariates <- "covariate_1 + covariate_2 + covariate_3"}

\item{effect_modifier}{A string specifying an effect modifier to interact with the exposure. Similar to "covariates"}

\item{sensitivity_cov}{A character vector of sensitivity covariates. Similar to "covariates"}

\item{random_effects}{A string specifying the random effects structure (e.g., "(1 | group)").}
}
\value{
A list containing the fitted model, a tidy summary table, the model formula, residuals, and exposure name.
}
\description{
This function fits a linear model (`lm`) or a mixed effects model (`lmer`) depending on the presence of random effects. It supports interaction terms, sensitivity covariates, and returns a tidy summary of the model.
}
