% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttl_statements_write.R
\name{ttl_statements_write}
\alias{ttl_statements_write}
\title{Write statements into a TTL file}
\usage{
ttl_statements_write(tdf, file_path, overwrite = TRUE)
}
\arguments{
\item{tdf}{A dataset in exactly three columns.}

\item{file_path}{The path to the file that should be written or appended.}

\item{overwrite}{If the file exists, overwrite it? Defaults to \code{TRUE}.}
}
\description{
Write statements into a TTL file
}
\examples{
tdf <- data.frame (s = c("eg:01","eg:02",  "eg:01", "eg:02", "eg:01" ),
                   p = c("a", "a", "eg-var:", "eg-var:", "rdfs:label"),
                   o = c("qb:Observation",
                         "qb:Observation",
                         "\"1\"^^<xs:decimal>",
                         "\"2\"^^<xs:decimal>", 
                         '"Example observation"'))

examplefile <- file.path(tempdir(), "ttl_dataset_write.ttl")

dataset_ttl_write(tdf=tdf, file_path = examplefile)

readLines(examplefile)
}
