% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttl_write.R
\name{ttl_write}
\alias{ttl_write}
\title{Write a dataset into Turtle serialisation}
\usage{
ttl_write(tdf, ttl_namespace = NULL, file_path, overwrite = TRUE)
}
\arguments{
\item{tdf}{A dataset in exactly three columns.}

\item{ttl_namespace}{The namespace definitions of the dataset.}

\item{file_path}{The path to the file that should be written or appended.}

\item{overwrite}{If the file exists, overwrite it? Defaults to \code{TRUE}.}
}
\value{
A text file with the prefix and the observation serialisations.
}
\description{
Write a dataset into a Turtle serialisation.
}
\examples{
tdf <- data.frame(
  s = c("eg:01", "eg:02", "eg:01", "eg:02", "eg:01"),
  p = c("a", "a", "eg-var:", "eg-var:", "rdfs:label"),
  o = c(
    "qb:Observation",
    "qb:Observation",
    "\"1\"^^<xs:decimal>",
    "\"2\"^^<xs:decimal>",
    '"Example observation"'
  )
)

examplefile <- file.path(tempdir(), "ttl_dataset_write.ttl")

dataset_ttl_write(tdf = tdf, file_path = examplefile)

readLines(examplefile)
}
