% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_version_warning}
\alias{check_version_warning}
\title{Check for Package Version Warning}
\usage{
check_version_warning(current_version, latest_version = "0.1.6")
}
\arguments{
\item{current_version}{The current version of the package, typically from `packageVersion("yourpkg")`.}

\item{latest_version}{A character string indicating the latest available version (default is `"0.1.5"`).}
}
\value{
No return value. Called for its side effect: printing a message to the console.
}
\description{
This utility function checks whether the current version of the package is older than a specified latest version.
If so, it prints a message encouraging the user to update.
}
\details{
This function is useful for notifying users when they may be using an outdated version of the package.
It is intended to be called at the beginning of user-facing functions.
}
\examples{
check_version_warning(packageVersion("turtle"))

}
